+++ src/main/scala/com/google/cloud/spark/bigquery/BigQueryUtil.scala
@@ -15,7 +15,8 @@
  */
 package com.google.cloud.spark.bigquery
 
-import com.google.cloud.bigquery.TableId
+import com.google.cloud.bigquery.{BigQueryError, BigQueryException, TableId}
+import com.google.cloud.http.BaseHttpServiceException.UNKNOWN_CODE
 
 import scala.util.matching.Regex
 
@@ -63,4 +64,9 @@ object BigQueryUtil {
         .map(p => TableId.of(p, actualDataset, table))
         .getOrElse(TableId.of(actualDataset, table))
   }
+
+  def noneIfEmpty(s: String): Option[String] =  Option(s).filterNot(_.trim.isEmpty)
+
+  def convertAndThrow(error: BigQueryError) : Unit =
+    throw new BigQueryException(UNKNOWN_CODE, error.getMessage, error)
 }